remote_server/requirements.txt
# gRPC 依赖
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0

# 图像处理
numpy>=1.20.0
Pillow>=8.0.0

# PyTorch - 云端推理必需
torch>=2.0.0
torchvision>=0.15.0

# transformers（模型加载）
transformers>=4.30.0

